+++ src/types/Profunctor.js
@@ -5,6 +5,11 @@ export type Profunctor<Static, InA, InB, OutA, OutB> = {
     (OutA) => OutB,
     (InA, Static) => OutA
   ) => (InB, Static) => OutB,
+  mapInOut: (
+    (InB) => InA,
+    (OutA) => OutB,
+    (InA, Static) => OutA
+  ) => (InB, Static) => OutB,
   mapIn: ((InB) => InA, (InA, Static) => OutA) => (InB, Static) => OutA,
 
   mapOut: ((OutA) => OutB, (InA, Static) => OutA) => (InA, Static) => OutB,